move storage to separate package
