+++ cmd/server/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"errors"
 	"fmt"
+	"github.com/mixailo/go-training-metrics/cmd/server/metrics_storage"
 	"net/http"
 	"strconv"
 	"strings"
@@ -90,34 +91,7 @@ const (
 	TypeGauge   string = "gauge"
 )
 
-type MemStorage struct {
-	gauges   map[string]float64
-	counters map[string]int64
-}
-
-func (m *MemStorage) UpdateGauge(name string, value float64) {
-	m.gauges[name] = value
-}
-
-func (m *MemStorage) UpdateCounter(name string, value int64) {
-	oldValue, ok := m.counters[name]
-	if !ok {
-		m.counters[name] = value
-	} else {
-		m.counters[name] = oldValue + value
-	}
-}
-
-type MetricsStorage interface {
-	UpdateCounter(name string, value int64)
-	UpdateGauge(name string, value float64)
-}
-
-func NewStorage() MetricsStorage {
-	return &MemStorage{gauges: make(map[string]float64), counters: make(map[string]int64)}
-}
-
-var storage = NewStorage()
+var storage = metrics_storage.NewStorage()
 
 func main() {
 	mux := http.NewServeMux()
